update freight and product double to decimal types for deafult 0.0m

+++ textbooks/cs14net10/ch12/180326/ModernWeb/Northwind.EntityModels.Sqlite/Product.cs
@@ -29,7 +29,8 @@ public partial class Product
     public string? QuantityPerUnit { get; set; }
 
     [Column(TypeName = "money")]
-    public double? UnitPrice { get; set; }
+    //public double? UnitPrice { get; set; }
+    public decimal? UnitPrice { get; set; }
 
     [Column(TypeName = "smallint")]
     public short? UnitsInStock { get; set; }
